python/pyproject.toml
[project]
name = "pdf-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
    dependencies = [
        "openai>=1.102.0",
        "pymupdf>=1.26.4",
        "pytest>=8.4.1",
        "google-genai>=0.2.0",
    ]
